refactor: Extract components
